Compiler the scanned source code to ByteCode Instructions using Pratt Parsing technique
**Description**

Previously, we added logic to scan through the source code and
produce the Tokens on-demand.

In this step, we implement the parser which will scan through the
list of tokens, and compiles it into a Byte Code Instruction.

We are using the Pratt's Parsing technique to implement the
compiliation of byte code instructions.