Add support for String using generic Obj Representation
Motivation

Since, String can grow and shrink, we allocate string in Heap unlike with
number, bool and nil where we added them to stack using the value.

Here, the value payload is a Obj* which point to data on heap.

Since, we needed a generic representation for data on heap,
we used obj.