Free Memory Objects allocated on Heap for Strings
Motivation

Last change introduced string operations like concatenation, which
involves allocating new memory on the heap, and ignores the part
strings that was allocated, resulting in memory leak. Hence,
we keep track of all the objects using linked list, and at the end,
we traverse them and free all of them.