Add support for Hash Table
** Description **

This commit adds support for Hash Table, and String interning.

Hash Table support adding, getting and deleting the elements from
the table.

Hash Function converts the String to a integer using the entire string.
Then we take Modulo with the array length to get the bucket.

Collion detection is done using Linear Probing which is an open
addressing algorithm.

Deleting the elements is not as straightforward, hence, for deleting
we just mark the entry as a Tombstone. This is to ensure other nodes
are still reachable. If we mark node empty and nil, then its makes
other nodes unreachable.

String interning is a way of maintaining a single copy of the string
then way all string points to same string object in the memory. This
way pointer equality and string comparision is the same.