Add support for declaring and reading Global Variables
**Description**

This commit adds support for declaring and reading global
variables.
- Declaring a new variable using a var statement.
- Accessing the value of a variable using an identifier expression.

Global Variable are late bound, that is they are
resolved dynamically at runtime. That helps to define things
recursively. As a result we can refer to something which is
defined later, we just need to ensure they are available when
invoked.

Example
---------
fun showVariable() {
  print global;
}

var global = "after";
showVariable();

Recursive Example
--------
fun isOdd(n) {
  if (n == 0) return false;
  return isEven(n - 1);
}

fun isEven(n) {
  if (n == 0) return true;
  return isOdd(n - 1);
}

**Notes**

Why push value after OP_GET_GLOBAL?

Similar to how

var a = expression();
where expression can 3 + 5 * 7, who value is kept in stack
so later we can assign it to a, when we encounter OP_SET_GLOBAL

we would the same for say
var a = 3 + 5 * 7;
var b = a;

Hence, when we read "a", we need to push it to stack

**Testing**

test.txt
---------
var beverage = "cafe au lait";
var breakfast = "beignets with " + beverage;
print breakfast;

Output
---------
❯ ./clox test.txt
== code ==
0000   1 OP_CONSTANT         1 'cafe au lait'
0002   | OP_DEFINE_GLOBAL    0 'beverage'
0004   2 OP_CONSTANT         3 'beignets with '
0006   | OP_GET_GLOBAL       4 'beverage'
0008   | OP_ADD
0009   | OP_DEFINE_GLOBAL    2 'breakfast'
0011   3 OP_GET_GLOBAL       5 'breakfast'
0013   | OP_PRINT
0014   4 OP_RETURN
beignets with cafe au lait

0000   1 OP_CONSTANT         1 'cafe au lait'
         [ cafe au lait ]
0002   | OP_DEFINE_GLOBAL    0 'beverage'

0004   2 OP_CONSTANT         3 'beignets with '
         [ beignets with  ]
0006   | OP_GET_GLOBAL       4 'beverage'
         [ beignets with  ][ cafe au lait ]
0008   | OP_ADD
         [ beignets with cafe au lait ]
0009   | OP_DEFINE_GLOBAL    2 'breakfast'

0011   3 OP_GET_GLOBAL       5 'breakfast'
         [ beignets with cafe au lait ]
0013   | OP_PRINT
beignets with cafe au lait

0014   4 OP_RETURN