Add local variable support
**Description**

Unlike global variables, where get add lexeme to constant table,
then emit instruction to parse the global variable during runtime,
and add it to the table. Here, for local variable, we use the
local stack to do get and set.

**Motivation**

Previously we used VM stack for the temporaries, now we can use
the stack to store the local variable, and variable lookup is
array indexing, which is fast.

We maintain a separate stack to store the variable names, whose
slot matches with the vm.stack's value. Since, we leave the value
on stack.

{
var a = 10;
var b = 15;
}

local var stack --> a , b
vm stack        --> 10, 15

after we endscope, we pop those values.

** Testing **

test.c
----------
{
var a = 3 + 5;
  {
    var b = 1 * 2;
    var c = b + 20;
  }
}

Output
----------
== code ==
0000   2 OP_CONSTANT         0 '3'
0002   | OP_CONSTANT         1 '5'
0004   | OP_ADD
0005   4 OP_CONSTANT         2 '1'
0007   | OP_CONSTANT         3 '2'
0009   | OP_MULTIPLY
0010   5 OP_GET_LOCAL        1
0012   | OP_CONSTANT         4 '20'
0014   | OP_ADD
0015   6 OP_POP
0016   | OP_POP
0017   7 OP_POP
0018   8 OP_RETURN

0000   2 OP_CONSTANT         0 '3'
         [ 3 ]
0002   | OP_CONSTANT         1 '5'
         [ 3 ][ 5 ]
0004   | OP_ADD
         [ 8 ]
0005   4 OP_CONSTANT         2 '1'
         [ 8 ][ 1 ]
0007   | OP_CONSTANT         3 '2'
         [ 8 ][ 1 ][ 2 ]
0009   | OP_MULTIPLY
         [ 8 ][ 2 ]
0010   5 OP_GET_LOCAL        1  // NOTE, here we get are getting value from slot "1"
         [ 8 ][ 2 ][ 2 ]
0012   | OP_CONSTANT         4 '20'
         [ 8 ][ 2 ][ 2 ][ 20 ]
0014   | OP_ADD
         [ 8 ][ 2 ][ 22 ]
0015   6 OP_POP
         [ 8 ][ 2 ]
0016   | OP_POP
         [ 8 ]
0017   7 OP_POP

0018   8 OP_RETURN