Wrap top level into a function, and add support for CallFrames
**Description**

This commit wraps all the high level code, which is imperative into
automatically defined function. And also, every function has a callframe
which basically contains all the local variables in that function,
and a return address from where it was called.

**Motivation**

When we execute any function, we would want to return back to the
previous called function. This should be fairly straightforward,
we just update instruction pointer to the INST we want to execute
after we return. However, top level code is not in a function, hence
to keep things simple, we wrap all the top level code into a function.
This way, we are just jumping from one func to another, where we just
update the frame (one frame per function), which contains the IP to
start executing.

**Testing**

❯ cat ../test.c
var i;
for (i = 10; i > 2; i = 0 ) {
print i;
}
print "hello world";

❯ ./clox ../test.c
== <script> ==
0000   1 OP_NIL
0001   | OP_DEFINE_GLOBAL    0 'i'
0003   2 OP_CONSTANT         2 '10'
0005   | OP_SET_GLOBAL       1 'i'
0007   | OP_POP
0008   | OP_GET_GLOBAL       3 'i'
0010   | OP_CONSTANT         4 '2'
0012   | OP_GREATER
0013   | OP_JUMP_IF_FALSE   13 -> 34
0016   | OP_POP
0017   | OP_JUMP            17 -> 28
0020   | OP_CONSTANT         6 '0'
0022   | OP_SET_GLOBAL       5 'i'
0024   | OP_POP
0025   | OP_LOOP            25 -> 8
0028   3 OP_GET_GLOBAL       7 'i'
0030   | OP_PRINT
0031   4 OP_LOOP            31 -> 20
0034   | OP_POP
0035   5 OP_CONSTANT         8 'hello world'
0037   | OP_PRINT
0038   7 OP_RETURN