Add Foreign Function Interface to support the native function
*Description**

Unline generic function, here functions native for langauge
used, we don't actually create a new call frame, and
update the stack pointer to run it. Instead we just call
the native function with the arguments.

We provide a foreign function interface, here defineNative
which takes in the "name" to be called from this language,
and also a function pointer to function using the native function.

And, we register this directly in the "globals" table just like
any global function which gets defined. Here, since we don't need
call frame, code etc, we don't have to worry about emitting
instructions to define in global table.