Add support for upvalue lookups in closure
**Description**

In closure if we look in the surrounding
function for value.
If found, then we add the index while performing
lexical scoping.

If not found, we add the index of the surronding
function's upvalue array.

fun outer() {
  var x = 1;
  fun middle() {
    fun inner() {
      print x;
    }
  }
}

Here, in inner(), x is not found, look in surrounding
function, its not declared there as well.

middle will have a upvalue capturing "x"
inner will have a upvalue which will capture middle's upvalue index.

Test.c
======
fun outer() {
  var a = 1;
  var b = 2;
  fun middle() {
    var c = 3;
    var d = 4;
    fun inner() {
      print a + c + b + d;
    }
  }
}

Output
=======

0004   9 OP_CLOSURE          2 <fn inner>
0006      |                     upvalue 0
0008      |                     local 1
0010      |                     upvalue 1
0012      |                     local 2