Mark the indirect reference Objects
**Description**

In the initial step, we scanned through the intial directly reachable
objects, called "roots".

We can add all these "root" objects to a worklist to loop
over and add its referenced objects, and indirected referenced obj's
reference. This is essentially traversing the graph to process these
"gray" object. Once we have process all the references to gray object
we mark it black.

All the nodes which we initially didn't mark are white nodes. Hence,
we can clean up the white nodes as there is no way to reach those.