Set Garbage Collector Run frequency using self adjusting heap
**Motivation**

How often do we want to run the GC?

We need to consider 2 factors?

Throughput: Throughput is the total fraction of time spent running user
code versus doing garbage collection work. Say you run a program for t
en seconds and it spends a second of that inside collectGarbage().
That means the throughput is 90%—it spent 90% of the time running
the program and 10% on GC overhead.

Latency: Latency is the longest continuous chunk of time where the user’s
program is completely paused while garbage collection happens. It’s a measure
of how “chunky” the collector is. Latency is an entirely different metric
than throughput. Its basically in the worst case scenario what's the longest
a program must wait before resuming execution.

- If we push running GC for a long time, then when we run the GC, we need
to collect more objects, process. Hence increased latency. But better
throughput.

- If we run GC frquently, user program should wait every few instance, before
resuming. Here "longest" wait is less. Better latency, but bad throughput.

Hence, we need to strike a balance. We use self adjusting heaps.

**Description**

Here, we essentially set the "after how many bytes we should run the
GC"?

So, after sweep based on the memory freed and memory left/bytes Consumed
we set the next GC run. If bytesAllocated left is high, we multiply by
a factor of 2. Hence, we further push the traversing if memory is high
optimizing for better throughput.

If bytesAllocated is less, multiply by 2, won't push it further away
we run GC frequenctly optimizing for latency.