Handle Garbage Collector Edge cases
**Motivation**

Since, we are marking all the roots, and traverse all its
indirectly references recursively, where do we not scan
or look for the object.

"Native Stack", here "C stack".

For instance, lets consider a scenario, where we are
adding a constant, it can be some value object. Now,
we create this "value object" and call our helper
function `addConstant()`, right now this object
only exist as function parameter (C stack), as a result,
If the constant table doesn't have the capacity, it will
trigger allocate which will not mark this object, as
a result it will be cleaned up. Hence, we temporarily push
it to stack, post Writing to constant table, we push it back.

**Description**

This commit handle several other edge cases like mentioned in the
Motivation Section.