Optimize the method call
**Description**

This commit optimizes the method call,
by skipping the intermediate bound object creation.

**Motivation**

Current semantics define a method invocation as two operations—accessing
the method and then calling the result. Our VM must support those as
separate operations because the user can separate them.

For instance:

class Brunch {
  eggs() {}
}

var brunch = Brunch();
var e = brunch.eggs;
e();

After the compiler has parsed the property name, we look for a left parenthesis.
If we match one, we switch to a new code path. There, we compile the argument list
exactly like we do when compiling a call expression. Then we emit a single new
OP_INVOKE instruction. It takes two operands:

1. The index of the property name in the constant table.

2. The number of arguments passed to the method.

In other words, this single instruction combines the operands of the OP_GET_PROPERTY
and OP_CALL instructions it replaces, in that order. It really is a fusion of those
two instructions.